Create LICENSE file
The README says "This plugin is released under the Unlicense. See the LICENSE file for details.", however there is no file with that name.  This adds the unlicense text from http://unlicense.org/ to a new LICENSE file.

Possible enhancements: add a link in the README to the LICENSE file, or renaming LICENSE to UNLICENSE per the recommendation at http://unlicense.org/.